+++ waton/__init__.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 from .core.errors import ConnectionError as ConnectionError  # noqa: A004
 
-__version__ = "0.1.4rc1"
+__version__ = "0.1.4rc3"
 
 
 __all__ = [